backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.1
sqlparse==0.5.3
psycopg2-binary==2.9.10
psycopg2-binary==2.9.10